+++ CHANGELOG.md
@@ -6,6 +6,7 @@ All notable changes to this project will be documented in this file.
 
 * Fixed: Set `max-width` for DB error logo image.
 * Fixed: Fix bug when checking table column exists for CPO.
+* Changed: Add error message when option Disable File Editor is activated.
 
 ## 1.5.0 - 2021-10-02